pyproject.toml
[tool.poetry]
name = "itask"
version = "0.1.0"
description = ""
authors = ["Hamid"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
flask-sqlalchemy = "^3.0.0"
Flask-Bcrypt = "^1.0.1"
Flask-Login = "^0.6.2"
Flask-RESTful = "^0.3.9"
rich = "^12.6.0"
flask-marshmallow = "^0.14.0"
jwt = "^1.3.1"
marshmallow-sqlalchemy = "^0.28.1"
PyJWT = "^2.5.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
